refactor(tangle-dapp): Separate rewards and points
