+++ apps/tangle-dapp/src/constants/reactQuery.ts
@@ -1,3 +1,4 @@
 export enum ReactQueryKey {
+  GetAccountRewards = 'GetAccountRewards',
   GetAccountPoints = 'GetAccountPoints',
 }
