--- apps/tangle-dapp/src/constants/swr.ts
@@ -1,3 +1,0 @@
-export enum SWRKey {
-  GetAccountRewards = 'GetAccountRewards',
-}
